@@ -10,11 +10,14 @@
     "url": "https://github.com/yuanyan/barium.git"
   },
   "dependencies": {
-    "react": "^0.12.0",
+    "react": "^0.14.4",
     "react-kit": "^0.1.3",
     "reactify": "^0.17.0"
   },
   "devDependencies": {
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "babelify": "^7.2.0",
     "browserify": "^6.3.3",
     "browserify-shim": "^3.8.0",
     "chalk": "^0.5.1",
