@@ -1,7 +1,9 @@
-var React = require('react');
-var converter = require('./converter');
-var hash = require('./hash');
+import React from 'react';
+import { rulesToString, ruleToString } from './converter';
+import hash from './hash';
+import getVendorPrefix from "react-kit/getVendorPrefix";
 
+const vendorPrefix = getVendorPrefix();
 var insertedRuleMap = {};
 var head = document.head || document.getElementsByTagName('head')[0];
 var styleTag;
@@ -21,25 +23,57 @@ function appendStyle(cssText) {
 
 }
 
-module.exports = {
-  create: function(styles) {
-    var cssText = '';
-    var ruleMap = {};
+export const create = (styles) => {
+  let cssText = '';
+  let ruleMap = {};
 
-    Object.keys(styles).forEach(function(val, key) {
-      var rules = styles[val];
-      var className = '_' + hash(JSON.stringify(rules)); // All with ._ prefix
-      var selector = '.' + className;
-      ruleMap[val] = className;
+  Object.keys(styles).forEach((val, key) => {
+    let rules = styles[val];
+    let className = `_${hash(JSON.stringify(rules))}`; // All with ._ prefix
+    let selector = `.${className}`;
+    ruleMap[val] = className;
 
-      if(!insertedRuleMap[selector]){
-        cssText += converter.rulesToString(selector, rules);
-      }
-      insertedRuleMap[selector] = true;
-    });
+    if(!insertedRuleMap[selector]){
+      cssText += rulesToString(selector, rules);
+    }
+    insertedRuleMap[selector] = true;
+  });
 
-    appendStyle(cssText);
+  appendStyle(cssText);
 
-    return ruleMap;
+  return ruleMap;
+};
+
+export const createKeyframes = (keyframes) => {
+  let animationName = `_${hash(JSON.stringify(keyframes))}`;
+  let cssText = `
+    @${vendorPrefix}keyframes ${animationName} {
+  `;
+
+  for (var frame in keyframes) {
+    cssText += `
+      ${frame} {
+    `;
+
+    for (var cssProperty in keyframes[frame]) {
+      let cssValue = keyframes[frame][cssProperty];
+      cssText += ruleToString(cssProperty, cssValue);
+    }
+
+    cssText += "}";
   }
+
+  cssText += "}";
+
+  appendStyle(cssText);
+
+  return animationName;
+};
+
+export const createAnimations = (animations) => {
+  let animationMap = {};
+  Object.keys(animations).forEach((animation) => {
+    animationMap[animation] = createKeyframes(animations[animation]);
+  });
+  return animationMap;
 }
