@@ -1,7 +1,3 @@
-function isValidValue(value) {
+export default (value) => {
   return value !== '' && (typeof value === 'number' || typeof value === 'string');
 }
-
-module.exports = {
-  isValidValue: isValidValue
-};
